requirements.txt
torch
Pillow
pandas
sqlalchemy

